feat: Update create-note use case to allow optional file attachment
This commit modifies the create-note use case to allow for an optional file attachment. The file parameter in the CreateNoteDTO interface is now marked as optional, enabling users to create notes without attaching a file. This change provides more flexibility and accommodates scenarios where a file attachment is not necessary.