chore: Update npm test scripts for use-cases and e2e tests
The npm test scripts in the package.json file have been updated to include separate commands for running use-case tests and end-to-end (e2e) tests. The "test:use-cases" script now runs all the tests in the "src/**/use-cases" directory, while the "test:e2e" script runs all the tests in the "src/**/controller" directory. This change improves the organization and clarity of the test scripts, making it easier to run specific types of tests.